Eliminated a couple compilation warnings in isoview.cpp regarding assignment of float to integer.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@6889 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/iso/isoview.cpp
@@ -257,7 +257,7 @@ void csIsoFakeCamera::IsoReady(const csVector3& position,
   shiftx = scroll.x + minz * trans.GetO2T().m13;
   shifty = scroll.y + minz * trans.GetO2T().m23;
 
-  rview->GetG3D()->SetPerspectiveCenter(shiftx, shifty);
+  rview->GetG3D()->SetPerspectiveCenter((int)shiftx, (int)shifty);
   rview->GetG3D()->SetClipper( rview->GetClipper(), CS_CLIPPER_TOPLEVEL);
 
   //csMatrix3 bb = trans.GetO2T();
